Move metrics module to a child package
