+++ metrics/resource/vm.go
@@ -1,12 +1,14 @@
-package minivmm
+package resource
 
 import (
 	"log"
 	"strconv"
+
+	"minivmm"
 )
 
-// Metric is the metrics of resrouces that minivmm manages. It's exported for prometheus.
-type Metric struct {
+// VMMetric is the metrics of resrouces that minivmm manages. It's exported for prometheus.
+type VMMetric struct {
 	CPUCores           int `json:"minivmm_cpu_cores"`
 	CPUCoresRunning    int `json:"minivmm_cpu_cores_running"`
 	MemoryBytes        int `json:"minivmm_memory_bytes"`
@@ -16,10 +18,11 @@ type Metric struct {
 	NumVMRunning       int `json:"minivmm_vms_running"`
 }
 
-func GetMetric() (*Metric, error) {
-	var m Metric
+// GetVMMetric returns the resource metrics of the VMs managed by minivmm .
+func GetVMMetric() (*VMMetric, error) {
+	var m VMMetric
 
-	vms, err := ListVMs()
+	vms, err := minivmm.ListVMs()
 	if err != nil {
 		return nil, err
 	}
@@ -27,8 +30,8 @@ func GetMetric() (*Metric, error) {
 	m.NumVM = len(vms)
 	for _, vm := range vms {
 		cpuStr := vm.CPU
-		memStr, _ := convertSIPrefixedValue(vm.Memory, "")
-		diskStr, _ := convertSIPrefixedValue(vm.Disk, "")
+		memStr, _ := minivmm.ConvertSIPrefixedValue(vm.Memory, "")
+		diskStr, _ := minivmm.ConvertSIPrefixedValue(vm.Disk, "")
 
 		cpu, err := strconv.Atoi(cpuStr)
 		if err != nil {
